added a comment and clean up
